debug Action

+++ test/test-mk-release-on-dev.sh
@@ -12,15 +12,23 @@ echo SCRIPTDIRPATH: "$SCRIPTDIRPATH"
 
 #patch mk-release.sh
 cp "$SCRIPTDIRPATH"/mk-release.sh /tmp/ || exit 1
-sed -i 's/"\^v\\d+/"^DEV\\d+/' /tmp/mk-release.sh
+sed -i 's/"\^v\\d+/"\^DEV\\d+/' /tmp/mk-release.sh
+echo patched TAG regex in /tmp/mk-release.sh:
+grep "TAG=" /tmp/mk-release.sh
+
 
 # latest DEV tag
 LATEST_DEV=$(git tag|grep -P  "DEV\d.\d"|sort|tail -n1)
-
+echo LATEST_DEV:  "$LATEST_DEV"
 # build a "release" based on latest DEV tag
 /tmp/mk-release.sh "$LATEST_DEV"
+if [[ "$?" != "0" ]]; then
+    echo a release tar ball was not produced, exiting
+    exit 1
+fi
 
-UNTAR_LOCATION=/tmp/dar-test-install/
+
+UNTAR_LOCATION=/tmp/dar-test-install/ 
 TAR_FILE=dar-backup-linux-${LATEST_DEV}.tar.gz
 echo TAR_FILE to test install of: "$TAR_FILE"
 rm -fr "$UNTAR_LOCATION" || exit 1
